+++ lib/sx126x_hal/radio_hal.cpp
@@ -11,14 +11,14 @@
 // ===================== DIO1 CAPTURE GLOBALS =====================
 
 volatile bool     dio1Fired      = false;
-volatile uint32_t dio1CaptureVal = 0;
+volatile uint64_t dio1CaptureVal = 0;
 
 // ===================== GPIO ISR =====================
 
 volatile uint32_t dio1IsrCount = 0;  // incremented every ISR fire; read from loop for diagnostics
 
 static void IRAM_ATTR dio1Isr() {
-    dio1CaptureVal = (uint32_t)micros();
+    dio1CaptureVal = micros();
     dio1Fired      = true;
     dio1IsrCount++;
 }
